passes: Move DeviceOmpizer to passes/iet
