+++ devito/core/gpu_openmp.py
@@ -1,288 +1,14 @@
 from functools import singledispatch
 
 import cgen as c
-from sympy import Function
 
-from devito.core.gpu import (DeviceNoopOperator, DeviceOperator, DeviceCustomOperator,
-                             is_on_device)
-from devito.data import FULL
-from devito.ir.equations import DummyEq
-from devito.ir.iet import (Block, Call, Conditional, EntryFunction,
-                           List, LocalExpression, FindNodes, ParallelTree,
-                           MapExprStmts, Transformer)  #TODO refactor
-from devito.mpi.distributed import MPICommObject
-from devito.mpi.routines import IrecvCall, IsendCall
-from devito.passes.iet import (DataManager, Ompizer, OpenMPIteration,
-                               Storage, iet_pass)  #TODO: TO BE MOVED...
-from devito.symbolics import CondNe, Byref, ccode
+from devito.core.gpu import DeviceNoopOperator, DeviceOperator, DeviceCustomOperator
+from devito.ir.iet import EntryFunction, MapExprStmts
+from devito.passes.iet import DataManager, DeviceOmpizer, Storage, iet_pass, is_on_device  #TODO: TO BE MOVED...
 from devito.tools import filter_sorted
-from devito.types import DeviceID, DeviceRM, Symbol
+from devito.types import DeviceRM
 
 __all__ = ['DeviceNoopOmpOperator', 'DeviceAdvOmpOperator', 'DeviceCustomOmpOperator']
-
-
-class DeviceOmpIteration(OpenMPIteration):
-
-    @classmethod
-    def _make_header(cls, **kwargs):
-        header, kwargs = super()._make_header(**kwargs)
-        kwargs.pop('gpu_fit', None)
-
-        return header, kwargs
-
-    @classmethod
-    def _make_construct(cls, **kwargs):
-        return 'omp target teams distribute parallel for'
-
-    @classmethod
-    def _make_clauses(cls, **kwargs):
-        kwargs['chunk_size'] = False
-        return super(DeviceOmpIteration, cls)._make_clauses(**kwargs)
-
-
-class DeviceOmpizer(Ompizer):
-
-    lang = dict(Ompizer.lang)
-    lang.update({
-        'map-enter-to': lambda i, j:
-            c.Pragma('omp target enter data map(to: %s%s)' % (i, j)),
-        'map-enter-alloc': lambda i, j:
-            c.Pragma('omp target enter data map(alloc: %s%s)' % (i, j)),
-        'map-update': lambda i, j:
-            c.Pragma('omp target update from(%s%s)' % (i, j)),
-        'map-update-host': lambda i, j:
-            c.Pragma('omp target update from(%s%s)' % (i, j)),
-        'map-update-device': lambda i, j:
-            c.Pragma('omp target update to(%s%s)' % (i, j)),
-        'map-release': lambda i, j, k:
-            c.Pragma('omp target exit data map(release: %s%s)%s'
-                     % (i, j, k)),
-        'map-exit-delete': lambda i, j, k:
-            c.Pragma('omp target exit data map(delete: %s%s)%s'
-                     % (i, j, k)),
-    })
-
-    _Iteration = DeviceOmpIteration
-
-    def __init__(self, sregistry, options, platform):
-        super().__init__(sregistry, options, platform)
-
-        self.gpu_fit = options['gpu-fit']
-        self.par_disabled = options['par-disabled']
-
-    @classmethod
-    def _make_sections_from_imask(cls, f, imask):
-        datasize = cls._map_data(f)
-        if imask is None:
-            imask = [FULL]*len(datasize)
-        assert len(imask) == len(datasize)
-        sections = []
-        for i, j in zip(imask, datasize):
-            if i is FULL:
-                start, size = 0, j
-            else:
-                try:
-                    start, size = i
-                except TypeError:
-                    start, size = i, 1
-                start = ccode(start)
-            sections.append('[%s:%s]' % (start, size))
-        return ''.join(sections)
-
-    @classmethod
-    def _map_data(cls, f):
-        if f.is_Array:
-            return f.symbolic_shape
-        else:
-            return tuple(f._C_get_field(FULL, d).size for d in f.dimensions)
-
-    @classmethod
-    def _map_to(cls, f, imask=None, queueid=None):
-        sections = cls._make_sections_from_imask(f, imask)
-        return cls.lang['map-enter-to'](f.name, sections)
-
-    _map_to_wait = _map_to
-
-    @classmethod
-    def _map_alloc(cls, f, imask=None):
-        sections = cls._make_sections_from_imask(f, imask)
-        return cls.lang['map-enter-alloc'](f.name, sections)
-
-    @classmethod
-    def _map_present(cls, f, imask=None):
-        return
-
-    @classmethod
-    def _map_update(cls, f):
-        return cls.lang['map-update'](f.name, ''.join('[0:%s]' % i
-                                                      for i in cls._map_data(f)))
-
-    @classmethod
-    def _map_update_host(cls, f, imask=None, queueid=None):
-        sections = cls._make_sections_from_imask(f, imask)
-        return cls.lang['map-update-host'](f.name, sections)
-
-    _map_update_wait_host = _map_update_host
-
-    @classmethod
-    def _map_update_device(cls, f, imask=None, queueid=None):
-        sections = cls._make_sections_from_imask(f, imask)
-        return cls.lang['map-update-device'](f.name, sections)
-
-    _map_update_wait_device = _map_update_device
-
-    @classmethod
-    def _map_release(cls, f, devicerm=None):
-        return cls.lang['map-release'](f.name,
-                                       ''.join('[0:%s]' % i for i in cls._map_data(f)),
-                                       (' if(%s)' % devicerm.name) if devicerm else '')
-
-    @classmethod
-    def _map_delete(cls, f, imask=None, devicerm=None):
-        sections = cls._make_sections_from_imask(f, imask)
-        # This ugly condition is to avoid a copy-back when, due to
-        # domain decomposition, the local size of a Function is 0, which
-        # would cause a crash
-        items = []
-        if devicerm is not None:
-            items.append(devicerm.name)
-        items.extend(['(%s != 0)' % i for i in cls._map_data(f)])
-        cond = ' if(%s)' % ' && '.join(items)
-        return cls.lang['map-exit-delete'](f.name, sections, cond)
-
-    @classmethod
-    def _map_pointers(cls, f):
-        raise NotImplementedError
-
-    def _make_threaded_prodders(self, partree):
-        if isinstance(partree.root, DeviceOmpIteration):
-            # no-op for now
-            return partree
-        else:
-            return super()._make_threaded_prodders(partree)
-
-    def _make_partree(self, candidates, nthreads=None):
-        """
-        Parallelize the `candidates` Iterations attaching suitable OpenMP pragmas
-        for parallelism. In particular:
-
-            * All parallel Iterations not *writing* to a host Function, that
-              is a Function `f` such that ``is_on_device(f) == False`, are offloaded
-              to the device.
-            * The remaining ones, that is those writing to a host Function,
-              are parallelized on the host.
-        """
-        assert candidates
-        root = candidates[0]
-
-        if is_on_device(root, self.gpu_fit, only_writes=True):
-            # The typical case: all written Functions are device Functions, that is
-            # they're mapped in the device memory. Then we offload `root` to the device
-
-            # Get the collapsable Iterations
-            collapsable = self._find_collapsable(root, candidates)
-            ncollapse = 1 + len(collapsable)
-
-            body = self._Iteration(gpu_fit=self.gpu_fit, ncollapse=ncollapse, **root.args)
-            partree = ParallelTree([], body, nthreads=nthreads)
-            collapsed = [partree] + collapsable
-
-            return root, partree, collapsed
-        elif not self.par_disabled:
-            # Resort to host parallelism
-            return super()._make_partree(candidates, nthreads)
-        else:
-            return root, None, None
-
-    def _make_parregion(self, partree, *args):
-        if isinstance(partree.root, DeviceOmpIteration):
-            # no-op for now
-            return partree
-        else:
-            return super()._make_parregion(partree, *args)
-
-    def _make_guard(self, parregion, *args):
-        partrees = FindNodes(ParallelTree).visit(parregion)
-        if any(isinstance(i.root, DeviceOmpIteration) for i in partrees):
-            # no-op for now
-            return parregion
-        else:
-            return super()._make_guard(parregion, *args)
-
-    def _make_nested_partree(self, partree):
-        if isinstance(partree.root, DeviceOmpIteration):
-            # no-op for now
-            return partree
-        else:
-            return super()._make_nested_partree(partree)
-
-    @iet_pass
-    def make_gpudirect(self, iet):
-        """
-        Modify MPI Callables to enable multiple GPUs performing GPU-Direct communication.
-        """
-        mapper = {}
-        for node in FindNodes((IsendCall, IrecvCall)).visit(iet):
-            header = c.Pragma('omp target data use_device_ptr(%s)' %
-                              node.arguments[0].name)
-            mapper[node] = Block(header=header, body=node)
-
-        iet = Transformer(mapper).visit(iet)
-
-        return iet, {}
-
-    @iet_pass
-    def initialize(self, iet):
-
-        @singledispatch
-        def _initialize(iet):
-            return iet, {}
-
-        @_initialize.register(EntryFunction)
-        def _(iet):
-            # TODO: we need to pick the rank from `comm_shm`, not `comm`,
-            # so that we have nranks == ngpus (as long as the user has launched
-            # the right number of MPI processes per node given the available
-            # number of GPUs per node)
-
-            objcomm = None
-            for i in iet.parameters:
-                if isinstance(i, MPICommObject):
-                    objcomm = i
-                    break
-
-            deviceid = DeviceID()
-            if objcomm is not None:
-                rank = Symbol(name='rank')
-                rank_decl = LocalExpression(DummyEq(rank, 0))
-                rank_init = Call('MPI_Comm_rank', [objcomm, Byref(rank)])
-
-                ngpus = Symbol(name='ngpus')
-                call = Function('omp_get_num_devices')()
-                ngpus_init = LocalExpression(DummyEq(ngpus, call))
-
-                osdd_then = Call('omp_set_default_device', [deviceid])
-                osdd_else = Call('omp_set_default_device', [rank % ngpus])
-
-                body = [Conditional(
-                    CondNe(deviceid, -1),
-                    osdd_then,
-                    List(body=[rank_decl, rank_init, ngpus_init, osdd_else]),
-                )]
-            else:
-                body = [Conditional(
-                    CondNe(deviceid, -1),
-                    Call('omp_set_default_device', [deviceid])
-                )]
-
-            init = List(header=c.Comment('Begin of OpenMP+MPI setup'),
-                        body=body,
-                        footer=(c.Comment('End of OpenMP+MPI setup'), c.Line()))
-            iet = iet._rebuild(body=(init,) + iet.body)
-
-            return iet, {'args': deviceid}
-
-        return _initialize(iet)
 
 
 class DeviceOmpDataManager(DataManager):
