+++ devito/passes/iet/__init__.py
@@ -4,4 +4,5 @@ from .mpi import *  # noqa
 from .misc import *  # noqa
 from .definitions import *  # noqa
 from .instrument import *  # noqa
+from .language import *  # noqa
 from .specializations import *  # noqa
